refactored whole code
