+++ src/hesu/hesuadmin.py
@@ -1,4 +1,4 @@
-from hesuutils import *
+from src.utils.hesuutils import *
 
 class Admin(object):
     def __init__(self, public_key, private_key):
