--- test.py
@@ -1,3 +1,0 @@
-import He_Su_impl
-
-He_Su_impl.vote()
