+++ test/test.py
@@ -1,0 +1,3 @@
+from src import hesuuser
+
+hesuuser.vote()
